@@ -15,7 +15,7 @@ publish_to: "none"
 version: 0.0.6+6
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.10.0 <3.0.0"
 
 dependencies:
   flutter:
