@@ -29,7 +29,7 @@ class _DriverHomeState extends State<DriverHome> {
     return BlocConsumer<ContainerBloc, ContainerState>(
       listener: (context, state) {
         if (state is ContainerFailure) {
-          Scaffold.of(context).showSnackBar(SnackBar(
+          ScaffoldMessenger.of(context).showSnackBar(SnackBar(
             content: Text(state.message),
           ));
         }
